update: changes multiplier and bonus fields to float64 type

+++ internal/adapters/http/dto/campaign_dto.go
@@ -1,13 +1,13 @@
 package dto
 
 type CampaignDTO struct {
-	ID                   string `json:"id"`
-	CommerceID           string `json:"commerce_id"`
-	BranchID             string `json:"branch_id"`
-	CampaignName         string `json:"campaign_name"`
-	CampaignType         string `json:"campaign_type"`
-	CampaignMultiplier   int32  `json:"campaign_multiplier"`
-	CampaignPercentBonus int16  `json:"campaign_percent_bonus"`
-	StartDate            string `json:"start_date"`
-	EndDate              string `json:"end_date"`
+	ID                   string  `json:"id"`
+	CommerceID           string  `json:"commerce_id"`
+	BranchID             string  `json:"branch_id"`
+	CampaignName         string  `json:"campaign_name"`
+	CampaignType         string  `json:"campaign_type"`
+	CampaignMultiplier   float64 `json:"campaign_multiplier"`
+	CampaignPercentBonus float64 `json:"campaign_percent_bonus"`
+	StartDate            string  `json:"start_date"`
+	EndDate              string  `json:"end_date"`
 }
